Moved component-model from compiler-arg to config
