+++ mapstruct-quarkus/pom.xml
@@ -91,9 +91,6 @@
               <version>${org.mapstruct.version}</version>
             </path>
           </annotationProcessorPaths>
-          <compilerArgs>
-            <compilerArg>-Amapstruct.defaultComponentModel=cdi</compilerArg>
-          </compilerArgs>
         </configuration>
       </plugin>
       <plugin>
